Making Entry Model swappable using django-swappable-models.
    * With this change, there is no need for the app extending
      Entry model to provide workarounds like MIGRATION_MODULES
      or ZINNIA_ENTRY_BASE_MODEL.
    * The app just needs to make a dependency in the migration
      to zinnia by adding

      ('zinnia', '0001_initial'),

      to depencies section in the generated migration. After that
      first zinnia:0001_initial migration is run, followed by
      the migration generated in the app extending the Entry model.
      And then rest of the migrations defined in zinnia/migrations/
      are run.

      This way the migrations are generated in the correct place.
