+++ buildout.cfg
@@ -25,6 +25,7 @@ zinnia                  = beautifulsoup4
                           mots-vides
                           pyparsing
                           regex
+                          swapper
 markups                 = docutils
                           markdown
                           textile
