+++ setup.py
@@ -32,7 +32,8 @@ setup(
     license=zinnia.__license__,
     include_package_data=True,
     zip_safe=False,
-    install_requires=['beautifulsoup4>=4.3.2',
+    install_requires=['swapper>=1.0.0',
+                      'beautifulsoup4>=4.3.2',
                       'django-contrib-comments>=1.7.2',
                       'django-mptt>=0.8.6',
                       'django-tagging>=0.4.5',
