+++ versions.cfg
@@ -43,5 +43,6 @@ sh                              = 1.12.9
 snowballstemmer                 = 1.2.1
 sphinx                          = 1.5.3
 sphinx-rtd-theme                = 0.1.9
+swapper                         = 1.0.0
 textile                         = 2.3.3
 zc.recipe.egg                   = 2.0.3
