+++ zinnia/comparison.py
@@ -9,12 +9,13 @@ from django.utils.functional import cached_property
 from django.utils.html import strip_tags
 
 import regex as re
+import swapper
 
-from zinnia.models.entry import Entry
 from zinnia.settings import COMPARISON_FIELDS
 from zinnia.settings import STOP_WORDS
 
 
+Entry = swapper.load_model("zinnia", "Entry")
 PUNCTUATION = re.compile(r'\p{P}+')
