+++ zinnia/managers.py
@@ -15,7 +15,9 @@ def tags_published():
     Return the published tags.
     """
     from tagging.models import Tag
-    from zinnia.models.entry import Entry
+    import swapper
+    Entry = swapper.load_model("zinnia", "Entry")
+
     tags_entry_published = Tag.objects.usage_for_queryset(
         Entry.published.all())
     # Need to do that until the issue #44 of django-tagging is fixed
