+++ zinnia/views/entries.py
@@ -1,13 +1,17 @@
 """Views for Zinnia entries"""
 from django.views.generic.dates import BaseDateDetailView
 
-from zinnia.models.entry import Entry
+import swapper
+
 from zinnia.views.mixins.archives import ArchiveMixin
 from zinnia.views.mixins.callable_queryset import CallableQuerysetMixin
 from zinnia.views.mixins.entry_cache import EntryCacheMixin
 from zinnia.views.mixins.entry_preview import EntryPreviewMixin
 from zinnia.views.mixins.entry_protection import EntryProtectionMixin
 from zinnia.views.mixins.templates import EntryArchiveTemplateResponseMixin
+
+
+Entry = swapper.load_model("zinnia", "Entry")
 
 
 class EntryDateDetail(ArchiveMixin,
