+++ zinnia/views/sitemap.py
@@ -1,9 +1,12 @@
 """Views for Zinnia sitemap"""
 from django.views.generic import TemplateView
 
+import swapper
+
 from zinnia.models.author import Author
 from zinnia.models.category import Category
-from zinnia.models.entry import Entry
+
+Entry = swapper.load_model("zinnia", "Entry")
 
 
 class Sitemap(TemplateView):
